i3plus: disable 2px curves for increased compatibility
The DMT48270M043_02WT LCD supports only 2 curves, instead of 4 for the
DMT48270M043_05WT LCD.  On the temperature graph page, 4 curves were being
used, two for bed temperature and two for extruder temperature, in order to
make the curves 2px high.  For compatibility with the _02WT model, just use
two curves.

Note that this requires new LCD resources, version 2.1.1.  Without this, on
the _05WT, this change will display both curves as red.  On the _02WT, the
behavior is undefined when 4 curves are present in the LCD resources screen.