i3plus: fix X homing when extruder is enabled
Due to the pinout of the ribbon cable from the mainboard to the interface
board on the gantry, there is significant noise present on both the GND and
the X_MIN pins.  GND is on the outside edge immediately next to an extruder
motor winding pin, and X_MIN is between two extruder motor winding pins.
Due to this, X_MIN is pulled up to ~4v9 at the limit switch with the
extruder off, but due to noise is only ~3v3 with the extruder enabled.  This
causes X_MIN to fall into an undefined range between V_IL and V_IH, but
seems to be sampled as a 0 (switch is closed), breaking homing.

Fix this by disabling the extruder motor before homing X.  While it could
possibly be expected to keep the extruder active while homing, there is not
really a valid use case.

It may also be possible to fix this issue by:
- Putting a stronger (external) pull-up on the X_MIN pin.
- Add a filtering capacitor to the X_MIN pin to filter stepper
- Splitting the X_MIN / GND lines into a separate shielded cable.
- Sample the pin using an ADC and use a lower threshold.  The response time
  may not be fast enough however.